+++ php/tests/Application/Command/ClassInfoTest/FinalClass.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace A;
+
+final class FinalClass
+{
+
+}