+++ php/tests/Application/Command/ClassInfoTest/FinalClassMethod.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace A;
+
+class TestClass
+{
+    final public function finalMethod()
+    {
+
+    }
+}